### 2021-05-26 #3317 @whedon
[PRE REVIEW]: RayFlare: flexible optical modelling of solar cells
**Submitting author:** @phoebe-p (<a href="http://orcid.org/0000-0001-9082-9506">Phoebe Pearce</a>)
**Repository:** <a href="https://github.com/qpv-research-group/rayflare" target ="_blank">https://github.com/qpv-research-group/rayflare</a>
**Version:** v1.0.0
