### 2021-07-01 #3317 @phoebe-p
Hi @arfon, my suggestions:
- @dalonsoa would be very suitable, but we have worked together on another software package related to RayFlare and he is also in the paper's acknowledgements (though he did not specifically help with the code for RayFlare), so perhaps this disqualifies him as a reviewer
- brandondube
- alvesjnr
- dsryberg
- samuelduchesne
- ivergara
