### 2021-07-06 #3317 @arfon
@dsryberg @EricaEgg @kanderso-nrel – would you be willing to review this submission for JOSS? The software under review is a Python package [_RayFlare: flexible optical modelling of solar cells_](https://github.com/qpv-research-group/rayflare).
