### 2021-07-06 #3317 @kanderso-nrel
I am willing to review but it may take longer than the (extended) 6-week turnaround -- I would want to get more familiar with optical modeling techniques before completing the review.  Happy to do it if nobody else is available to do it quicker.

Side note @phoebe-p, the `RayFlare_theory.pdf` document is amazing, very excited to read it!
