### 2021-07-06 #3317 @phoebe-p
@kanderso-nrel thank you! If you do read it, do let me know if you find anything which doesn't seem to make sense - it's a slightly adapted version of a chapter from my thesis so there may be parts which don't make sense/reference something which isn't actually in the document without the context of the rest of the thesis that I missed when adapting it.
