### 2021-07-07 #3317 @EricaEgg
I would also be happy to review.

Thanks for reaching out,

Erica Eggleton

On Tue, Jul 6, 2021 at 11:29 PM Arfon Smith ***@***.***>
wrote:

> I am willing to review but it may take longer than the (extended) 6-week
> turnaround -- I would want to get more familiar with optical modeling
> techniques before completing the review. Happy to do it if nobody else is
> available to do it quicker.
>
> Thanks @kanderso-nrel <https://github.com/kanderso-nrel> - that sounds
> great! I won't open the main review until I've identified a second reviewer
> here.
>
