### 2021-07-08 #3460 @whedon
[REVIEW]: RayFlare: flexible optical modelling of solar cells
**Submitting author:** @phoebe-p (<a href="http://orcid.org/0000-0001-9082-9506">Phoebe Pearce</a>)
**Repository:** <a href="https://github.com/qpv-research-group/rayflare" target ="_blank">https://github.com/qpv-research-group/rayflare</a>
**Version:** v1.0.0
