### 2021-07-08 #3460 @arfon
@kanderso-nrel, @EricaEgg - This is the review thread for the paper. All of our communications will happen here from now on.
