### 2021-08-05 #3460 @kanderso-nrel
@arfon I'm not sure the `LICENSE.txt` is acceptable as I don't think it matches the official GPLv3 text.  But I'm not a lawyer, can you take a look from JOSS's perspective?  Also @phoebe-p I notice that `setup.py` says the package is licensed LGPL rather than GPL, I suppose that should be made consistent: https://github.com/qpv-research-group/rayflare/blob/devel/setup.py#L74
