### 2021-08-05 #3460 @danielskatz
The author should take a look at the "How to apply this license" section of https://choosealicense.com/licenses/gpl-3.0/
