### 2021-08-05 qpv-research-group/rayflare#29 @kandersolar
[JOSS] add community guidelines
One of the JOSS review items is:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I don't think any of these three points are explicitly addressed (though maybe I missed it!).  Some packages include this sort of info in the GitHub README or as a page in their sphinx docs.  GitHub also has a feature for this that might be useful, see "Contributing" here: https://github.com/qpv-research-group/rayflare/community

xref https://github.com/openjournals/joss-reviews/issues/3460
