### 2021-08-05 qpv-research-group/rayflare#30 @kandersolar
[JOSS] paper.md formatting issue
There is a minor rendering issue in the JOSS manuscript on the last bullet point here:

![image](https://user-images.githubusercontent.com/57452607/128427141-e35e18ba-d741-4a55-83da-733be7ccb517.png)

I think the issue is that the source file has two dash characters instead of one for the "Ideal/theoretical reference cases" item.

xref https://github.com/openjournals/joss-reviews/issues/3460
