### 2021-08-05 qpv-research-group/rayflare#31 @kandersolar
[JOSS] Documentation (RTD) issues
Here is a list of the issues I noticed while browsing the documentation.  All of them are pretty minor IMHO.

## Index
- ~The bottom of `index.rst` is supposed to show the poster, but it doesn't render correctly on RTD:~ probably just some cache/loading issue on my first load; a hard refresh fixed it
![image](https://user-images.githubusercontent.com/57452607/128422303-4bdf6d69-e9a7-4118-90fd-6508935a4603.png)

## Examples
- Examples/rcwa_tmm_validation.html: figure titles are labeled "deg" but I think are actually in radians
- Examples/compare_models_3Jsolarcell.html: The figure labels are overlapping, consider using `(sharex=True, sharey=True)` or `tight_layout()`

## Function/Class docs
- matrix_formalism.process_structure.process_structure: duplicate "options" in docstring
- matrix_formalism.multiply_matrices.make_v0: phi_sym not documented
- matrix_formalism.ideal_cases.lambertian_matrix: structpath not documented, extra "options"
- matrix_formalism.ideal_cases.mirror_matrix: structpath not documented
- transfer_matrix_method.lookup_table: structpath not documented
- rigorous_coupled_wave_analysis.rcwa.get_reciprocal_lattice: sphinx rendering issue
- rigorous_coupled_wave_analysis.rcwa.rcwa_structure: in constructor and methods, "options" not documented
- rigorous_coupled_wave_analysis.rcwa.rcwa_structure.get_fields_z_integral: pol is not an actual parameter
- analytic module is not documented
- textures module is not documented
- various functions don't have docstrings

xref https://github.com/openjournals/joss-reviews/issues/3460
