### 2021-08-07 qpv-research-group/rayflare#32 @kandersolar
[JOSS] low and variable test coverage when running locally
This is more of a question than an issue report.  When I run the test suite locally and calculate coverage, I get a much lower number than what the CI finds.  The biggest coverage differences are `rt.py` and `rcwa.py`.  The specific numbers also change from run to run (e.g. in another run, `tmm.py` showed 99% coverage).  I'm happy to dig a bit more to try to understand why, but I wondered if this behavior is expected.  Here's the result of `$ coverage html`: [htmlcov.zip](https://github.com/qpv-research-group/rayflare/files/6949331/htmlcov.zip), and the corresponding test output:

```
$ pytest --cov-report= --cov=rayflare tests/

============================================= test session starts ==============================================
rootdir: /home/kevin/projects/rayflare
plugins: cov-2.12.1
collected 33 items

tests/test_analytic.py .                                                                                 [  3%]
tests/test_compare_methods.py ............                                                               [ 39%]
tests/test_ideal_cases.py ...                                                                            [ 48%]
tests/test_matrix_formalism.py .                                                                         [ 51%]
tests/test_ray_tracing.py .                                                                              [ 54%]
tests/test_rigorous_coupled_wave.py ......                                                               [ 72%]
tests/test_textures.py ....                                                                              [ 84%]
tests/test_transfer_matrix_method.py .....                                                               [100%]

<warning summary removed>

$ coverage report
Name                                                  Stmts   Miss  Cover
-------------------------------------------------------------------------
rayflare/__init__.py                                      0      0   100%
rayflare/analytic/__init__.py                             0      0   100%
rayflare/analytic/diffraction.py                         63     22    65%
rayflare/angles.py                                       57     28    51%
rayflare/matrix_formalism/__init__.py                     2      0   100%
rayflare/matrix_formalism/ideal_cases.py                 48      4    92%
rayflare/matrix_formalism/multiply_matrices.py          261      7    97%
rayflare/matrix_formalism/process_structure.py           97      5    95%
rayflare/options.py                                      31      0   100%
rayflare/ray_tracing/__init__.py                          1      0   100%
rayflare/ray_tracing/rt.py                              613    484    21%
rayflare/rigorous_coupled_wave_analysis/__init__.py       1      0   100%
rayflare/rigorous_coupled_wave_analysis/rcwa.py         621    336    46%
rayflare/state.py                                        12      3    75%
rayflare/structure.py                                    40      5    88%
rayflare/textures/__init__.py                             2      0   100%
rayflare/textures/define_textures.py                     16      0   100%
rayflare/textures/standard_rt_textures.py                58     12    79%
rayflare/transfer_matrix_method/__init__.py               1      0   100%
rayflare/transfer_matrix_method/lookup_table.py          74     61    18%
rayflare/transfer_matrix_method/tmm.py                  294    146    50%
rayflare/utilities.py                                    24      4    83%
-------------------------------------------------------------------------
TOTAL                                                  2316   1117    52%
```

xref https://github.com/openjournals/joss-reviews/issues/3460
