### 2021-08-09 qpv-research-group/rayflare#32 @phoebe-p
This is strange. I guess different test suites have slightly different ways of measuring coverage, and the 91% value that is reported on the RayFlare repository is calculated by codecov and not pytest, but the difference in total coverage (91 vs. 52%) seems very extreme, especially since codecov is calculating the coverage based on the test report generated by pytest. The fact that the coverage fluctuates is even stranger. The ray-tracing tests are obviously stochastic in nature but I can't think of a reason that would change the actual lines of code that get hit or not hit. Some tests have inputs which are randomly chosen from a range but again I cannot think of a reason why that would change which parts of the code are accessed. So basically, I have no idea why this is happening but I will try running pytest coverage locally on my own computer to see if I have the same problem.
