### 2021-08-09 qpv-research-group/rayflare#32 @phoebe-p
I have just had a thought on why the RCWA coverage could be so different -- what OS are you running this on? For the GitHub Actions, because installing S4 is a pain and easiest on Linux, I am only installing it and running the tests associated with RCWA/S4 on Ubuntu, so they are marked to be skipped if the operating system is not Linux:

```
@mark.skipif(sys.platform != "linux", reason="S4 (RCWA) only installed for tests under Linux")
```

So if you delete these lines where they appear, it should run on your system (assuming S4 is installed).

I need to check to be sure but this may well also affect coverage of TMM and RT-related files, since there are some examples which compare the different methods to see if they give consistent results, and if a test compares e.g. RCWA and TMM it will currently only run on Ubuntu so could also affect TMM coverage. Obviously, this isn't very good practice for writing tests, I should really have more unit tests which isolate these things and test smaller bits of code separately, but it might explain part of the issue, though it does not explain the fact that the coverage changes.
