### 2021-08-09 qpv-research-group/rayflare#32 @kandersolar
Ah sorry, I should have mentioned that I was running it on a linux (debian) installation with S4 installed.  If `skipif` was the culprit here I think we'd see `s` characters instead of `.` in the pytest output, so I don't think that's the issue, but I'll double-check that all the test functions are really executing as expected.  Could be that my virtual environment was messed up somehow, so I'll also start a fresh environment and see if that fixes things.  Hopefully if nothing else, I should be able to follow the same steps that the CI performs and track down the problem, but it would still be interesting to see if you can replicate the issue locally or not.
