### 2021-08-09 #3460 @phoebe-p
@kanderso-nrel I have just had a look at the 'How to Apply These Terms to Your New Programs' section of Daniel's link, and as far as I can tell the text I have in LICENSE.txt is the same as what is given there, though it is missing contact information as suggested and I do not currently have the license or a pointer to it in every individual file or terminal output, as is suggested by this section. I could add these things, have a copyright thing on every file is probably a good idea. Is there another version of this text/something else I am missing here?

Thanks for pointing out the inconsistency with setup.py, I will change that (I want to use GPL, not LGPL)!
