### 2021-08-09 #3460 @kanderso-nrel
Oh interesting @phoebe-p.  I had the impression that the LICENSE.txt is supposed to contain the full license text (several hundred lines long in the case of GPL) rather than that abbreviated notice.  My interpretation is that the abbreviated notice is intended for the top of source files, and when it says "You should have received a copy of the GNU General Public License along with this program", the source code file is the "program" and LICENSE.txt is the "copy" of the GPL that goes along with the source code.  But I'm very much not a lawyer, thus the appeal to JOSS for guidance :)
