### 2021-08-09 #3460 @phoebe-p
From some more research on the GNU website, it seems that it is indeed recommended you distribute a copy of the full license text with the software, so I will add this, since I see no reason not to. Either way, the 'How to apply this' section is clear that I am supposed to put a copyright notice at the start of every file, so I will also add that.
