### 2021-08-09 996b9f9
fix license and add copyright notice to start of each revelant file
