+++ LICENSE.txt
@@ -1,6 +1,6 @@
 RayFlare is a programme for flexible modelling of complicated optical structures.
 
-Copyright (C) 2020 Phoebe Pearce
+Copyright (C) 2021 Phoebe Pearce
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -14,4 +14,7 @@ GNU General Public License for more details.
 
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
-Also add information on how to contact you by electronic and paper mail.
+
+The full license text can be found in GNU_GPL_v3.txt
+
+Contact: pmp31@cam.ac.uk
