### 2021-08-09 qpv-research-group/rayflare#32 @kandersolar
Speculation: perhaps it has something to do with multiprocessing and lines executed by worker processes not getting counted correctly.  That could definitely make things behave differently run-to-run and between machines...
