### 2021-08-09 qpv-research-group/rayflare#32 @phoebe-p
Yes, good thinking. It also makes sense that in that case files which are covered separately (e.g. analytic/diffraction.py and angles.py) are consistent between my local run and the codecov report while the actual optical calculation parts which can happen in parallel are not. I will just make a new branch locally and switch all the calculations to not in parallel - this will definitely change the coverage somewhat, because obviously then the parts of the code for parallel execution will not be tested, but should give a more similar overall number if that is the cause.
