### 2021-08-10 qpv-research-group/rayflare#31 @phoebe-p
Re: the first issue, for me it looks like this? Could this just have been a loading issue? I'm looking at https://rayflare.readthedocs.io/en/latest/, were you at a different URL?
![Screenshot from 2021-08-10 14-54-19](https://user-images.githubusercontent.com/25822065/128884186-4acd35f2-f507-442d-9a3c-1a6262cc4ef6.png)
