### 2021-08-10 qpv-research-group/rayflare#31 @kandersolar
Strange -- clicking that link still showed me the `_images/poster.png` text instead of the image, but when I hard-refreshed the page the image appeared.  I'm willing to chalk it up to browser strangeness, so I'll cross it off the original list.
