### 2021-08-10 qpv-research-group/rayflare#31 @phoebe-p
To-do list for my own use:

- [x] Examples/rcwa_tmm_validation.html: figure titles are labeled "deg" but I think are actually in radians
- [x] Examples/compare_models_3Jsolarcell.html: The figure labels are overlapping, consider using (sharex=True, sharey=True) or tight_layout()

- [x] matrix_formalism.process_structure.process_structure: duplicate "options" in docstring
- [x] matrix_formalism.multiply_matrices.make_v0: phi_sym not documented
- [x] matrix_formalism.ideal_cases.lambertian_matrix: structpath not documented, extra "options"
- [x] matrix_formalism.ideal_cases.mirror_matrix: structpath not documented
- [x] transfer_matrix_method.lookup_table: structpath not documented
- [x] rigorous_coupled_wave_analysis.rcwa.get_reciprocal_lattice: sphinx rendering issue
- [x] rigorous_coupled_wave_analysis.rcwa.rcwa_structure: in constructor and methods, "options" not documented
- [x] rigorous_coupled_wave_analysis.rcwa.rcwa_structure.get_fields_z_integral: pol is not an actual parameter
- [x] analytic module is not documented
- [x] textures module is not documented
- [ ] various functions don't have docstrings
