### 2021-08-10 qpv-research-group/rayflare#29 @phoebe-p
To-do list for my own use:
- [x] Code of conduct
- [x] Guide for contributing
- [x] Issue template
- [x] Support information
