### 2021-08-10 68f92c3
:memo: add CONTRIBUTING.md
