+++ CONTRIBUTING.md
@@ -1,0 +1,136 @@
+# Contributing to RayFlare
+
+Thanks for your interest in contributing to RayFlare! The following is a set of guidelines for contributing to RayFlare, 
+an open-source, integrated optical modelling of complex stacks. The goal of these 
+guidelines is to make the development of the package efficient and sustainable and to ensure that every commit makes it 
+better, more readable, more robust and better documented. Please feel free suggest changes and improvements. 
+
+(this guide is based on the [Atom editor guide](https://github.com/atom/atom/blob/master/CONTRIBUTING.md))
+
+#### Table Of Contents
+
+[Code of Conduct](#code-of-conduct)
+
+[How Can I Contribute?](#how-can-i-contribute)
+  * [Reporting Bugs](#reporting-bugs)
+  * [Suggesting Enhancements](#suggesting-enhancements)
+  * [Pull Requests](#pull-requests)
+
+[Styleguides](#styleguides)
+  * [Git Commit Messages](#git-commit-messages)
+  * [Documentation Styleguide](#documentation-styleguide)
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by the [RayFlare Code of Conduct](CODE_OF_CONDUCT.md). 
+By participating, you are expected to uphold this code. Please report unacceptable behavior to 
+[the repository Administrator](mailto:pmp31@cam.ac.uk).
+
+## How Can I Contribute?
+
+### Reporting Bugs
+
+This section guides you through submitting a bug report for RayFlare. Following these guidelines helps maintainers and the 
+community understand your report, reproduce the behavior, and find related reports.
+
+Before creating bug reports, please check [this list](https://github.com/qpv-research-group/rayflare/issues) (including the closed 
+issues) as you might find out that you don't need to create one. When you are creating a bug report, please [include as many details as 
+possible](#how-do-i-submit-a-good-bug-report). 
+
+> **Note:** If you find a **Closed** issue that seems like it is the same thing that you're experiencing, open a new issue 
+>and include a link to the original issue in the body of your new one.
+
+#### How Do I Submit A (Good) Bug Report?
+
+Bugs are tracked as [GitHub issues](https://guides.github.com/features/issues/). Explain the problem and include additional 
+details to help maintainers reproduce the problem:
+
+* **Use a clear and descriptive title** for the issue to identify the problem.
+* **Describe the exact steps which reproduce the problem** in as many details as possible. For example, start by explaining how you installed RayFlare and what you where trying to do.
+* **Provide specific examples to demonstrate the steps**. Include links to files or GitHub projects, or copy/pasteable snippets, which you use in those examples. If you're providing snippets in the issue, use [Markdown code blocks](https://help.github.com/articles/markdown-basics/#multiple-lines).
+* **Describe the behavior you observed after following the steps** and point out what exactly is the problem with that behavior.
+* **Explain which behavior you expected to see instead and why.**
+* **If there is any error output in the temrinal, include that output with your report.**
+
+Provide more context by answering these questions:
+
+* **Did the problem start happening recently** (e.g. after updating to a new version of RayFlare) or was this always a problem?
+* If the problem started happening recently, **can you reproduce the problem in an older version of RayFlare?** What's the most recent version in which the problem doesn't happen? You can download older versions of RayFlare from [the releases page](https://github.com/qpv-research-group/rayflare/releases).
+* **Can you reliably reproduce the issue?** If not, provide details about how often the problem happens and under which conditions it normally happens.
+
+Include details about your configuration and environment:
+
+* **Which version of RayFlare are you using?** 
+* **What's the name and version of the OS you're using**?
+* **Are you running RayFlare in a virtual machine?** If so, which VM software are you using and which operating systems and versions are used for the host and the guest?
[…]
