### 2021-08-10 8c8735c
:memo: updates/fixes to docstrings
