### 2021-08-10 639d8ab
:memo: more documentation updates, added pages for Analytic and Texture modules
