+++ docs/Analytic/analytic.rst
@@ -1,0 +1,9 @@
+Analytical calculations
+==========================
+
+Analytic calculations relating to diffraction (i.e. not using S4).
+
+.. automodule:: rayflare.analytic.diffraction
+    :members:
+    :undoc-members:
+
