### 2021-08-10 d851198
codacy changes, enable plotting in tests again
