### 2021-08-10 qpv-research-group/rayflare#32 @phoebe-p
Ok, changing everything to run not in parallel improves the coverage somewhat (50% to 60%), but actually checking the coverage reports indicates that there is a more fundamental issue. For the rt.py, tmm.py and rcwa.py files, the coverage reports indicate that there is no coverage whatsoever of the RT, TMM and RCWA functions (i.e. the functions which generate matrices for the angular redistribution matrix method). This is incorrect, because these methods are used in several of the examples in test_compare_methods. The only thing I can think of is that codecov considers a part of the code covered if it is used while running the tests, and thus considers these functions (mostly) covered, while pytest imposes some more strict criterion, like that the function has to be called directly (which RCWA, TMM and RT are not -- they are called by process_structure in the test functions).
