### 2021-08-10 qpv-research-group/rayflare#29 @phoebe-p
@kanderso-nrel I think I have now covered all these things in PR #33, now merged into devel.
