+++ rayflare/textures/standard_rt_textures.py
@@ -11,12 +11,13 @@ import numpy as np
 import os
 
 def regular_pyramids(elevation_angle=55, upright=True, size=1):
-    """
+    """Defines RTSurface textures for ray-tracing of regular upright or inverted pyramids.
 
-    :param elevation_angle:
-    :param upright:
-    :param size:
-    :return:
+    :param elevation_angle: angle between the horizontal and a face of the pyramid, in degrees
+    :param upright: if True, upright pyramids. If False, inverted pyramids
+    :param size: size of the pyramids; the units are arbitrary, but should be kept consistent across
+            different interfaces if you are not randomizing the ray positions.
+    :return: a list of two RTSurface objects: [front_incidence, rear_incidence]
     """
 
     char_angle = math.radians(elevation_angle)
@@ -41,10 +42,10 @@ def regular_pyramids(elevation_angle=55, upright=True, size=1):
     return [surf_fi, surf_ri]
 
 def planar_surface(size=1):
-    """
+    """Defines RTSurface textures for ray-tracing for a planar surface for ray-tracing.
 
-    :param size:
-    :return:
+    :param size: size of the unit cell (this should not affect the results, as the surface is planar).
+    :return: a list of two RTSurface objects: [front_incidence, rear_incidence]
     """
     Lx = 1*size
     Ly = 1*size
@@ -60,9 +61,10 @@ def planar_surface(size=1):
 
 
 def random_pyramids():
-    """
+    """Defines RTSurface textures for ray-tracing for a surface of random pyramids (based on
+    real surface scan data).
 
-    :return:
+    :return: a list of two RTSurface objects: [front_incidence, rear_incidence]
     """
     cur_path = os.path.dirname(os.path.abspath(__file__))
     z_map = np.loadtxt(os.path.join(cur_path, 'pyramids.csv'), delimiter=',')
@@ -83,12 +85,13 @@ def random_pyramids():
 
 
 def V_grooves(elevation_angle=55, width=1, direction='y'):
-    """
+    """Defines RTSurface textures for ray-tracing for a surface of V-grooves.
 
-    :param elevation_angle:
-    :param width:
-    :param direction:
-    :return:
+    :param elevation_angle: angle between the horizontal and a face of the V-grooves, in degrres
+    :param width: width of the V-grooves (units are arbitrary but should be kept consistent between surfaces
+            if you are not randomizing the ray positions).
+    :param direction: Whether the V-grooves lie along the 'x' or 'y' direction (string)
+    :return: a list of two RTSurface objects: [front_incidence, rear_incidence]
     """
     char_angle = math.radians(elevation_angle)
     h = width*math.tan(char_angle)/2
