### 2021-08-10 qpv-research-group/rayflare#31 @phoebe-p
@kanderso-nrel I think these issues, apart from the final one (various functions don't have docstrings, i.e. the hardest one), were resolved by PR #33 + some subsequent changes to devel to update the documentation further. Currently, I think all the functions a user would actually want to call directly (in addition to many of the more 'internal' functions) have docstrings. Obviously, in the long term it would be good for all functions to have docstrings since it will help people who want to understand the code/contribute.
