### 2021-08-10 qpv-research-group/rayflare#30 @phoebe-p
Fixed by #33.
