### 2021-08-11 #3460 @phoebe-p
License issue has been fixed/updated; I've changed setup.py, added a full copy of the license in addition to the shorter statement, and added a copyright statement to each file.
