### 2021-08-11 qpv-research-group/rayflare#31 @kandersolar
Okay, makes sense!  If there's not a good use case for a user to call those functions themselves, one option would be to make them private by putting an underscore at the front of the function name.  I think that will keep them from showing up in the sphinx docs, as well as in IDE autocomplete suggestions and such.  NBD though, ok with me to consider this issue addressed and close it
