### 2021-08-11 qpv-research-group/rayflare#29 @kandersolar
I agree, feel free to close!
