### 2021-08-11 qpv-research-group/rayflare#31 @phoebe-p
Ah that's a good idea. At least for now this probably makes sense to avoid loads of undocumented functions showing up in the Sphinx docs.
