### 2021-08-13 qpv-research-group/rayflare#32 @kandersolar
I've made some partial progress here.  I added a couple files in the repository root:

```
$ cat .coveragerc
[run]
parallel = True
source = rayflare
$ cat sitecustomize.py
try:
    import coverage
    coverage.process_startup()
except ImportError:
    pass
```

- The `.coveragerc` file is needed for the `parallel=True` option, which means each subprocess will generate its own coverage file with a unique filename instead of all subprocesses using the same filename and clobbering each others' results.
- The `sitecustomize.py` file is needed so that coverage is enabled in the subprocesses in addition to the parent process.

Then I calculate coverage with this sequence:
- Generate coverage files: `PYTHONPATH=. COVERAGE_PROCESS_START=.coveragerc coverage run -m pytest`
    + Setting the environment variables is very important: `PYTHONPATH` is needed for `sitecustomize.py` to be recognized and executed; `COVERAGE_PROCESS_START` is needed for `coverage.process_startup()` to do anything.
- Merge coverage files: `coverage combine`
- Summarize: `coverage report`

```python
Name                                                  Stmts   Miss  Cover
-------------------------------------------------------------------------
rayflare/__init__.py                                      0      0   100%
rayflare/analytic/__init__.py                             0      0   100%
rayflare/analytic/diffraction.py                         63     22    65%
rayflare/angles.py                                       57     28    51%
rayflare/matrix_formalism/__init__.py                     2      0   100%
rayflare/matrix_formalism/ideal_cases.py                 48      4    92%
rayflare/matrix_formalism/multiply_matrices.py          261      7    97%
rayflare/matrix_formalism/process_structure.py           97      4    96%
rayflare/options.py                                      31      0   100%
rayflare/ray_tracing/__init__.py                          1      0   100%
rayflare/ray_tracing/rt.py                              613    139    77%
rayflare/rigorous_coupled_wave_analysis/__init__.py       1      0   100%
rayflare/rigorous_coupled_wave_analysis/rcwa.py         621    258    58%
rayflare/state.py                                        12      3    75%
rayflare/structure.py                                    40      5    88%
rayflare/textures/__init__.py                             2      0   100%
rayflare/textures/define_textures.py                     16      0   100%
rayflare/textures/standard_rt_textures.py                58     12    79%
rayflare/transfer_matrix_method/__init__.py               1      0   100%
rayflare/transfer_matrix_method/lookup_table.py          74      9    88%
rayflare/transfer_matrix_method/tmm.py                  294    145    51%
rayflare/utilities.py                                    24      4    83%
-------------------------------------------------------------------------
TOTAL                                                  2316    640    72%
```
So there's progress, but still lower than what codecov is finding.  Here's the full coverage report: [htmlcov.zip](https://github.com/qpv-research-group/rayflare/files/6985484/htmlcov.zip).  @phoebe-p maybe you can take a look at this report (or follow the above steps yourself) and see what you think?  Hopefully it should be more accurate about what is actually getting run.

I think this is beyond the scope of the JOSS review at this point -- I'm happy to continue helping figure this out, but I don't think there's any need to hold up the review for it.
