### 2021-08-22 #3460 @kanderso-nrel
I've completed the reviewer checklist and am happy to recommend this ms for acceptance.  Note that I do not consider https://github.com/qpv-research-group/rayflare/issues/32 to be a blocker for this review.  Please let me know if you need anything else from me.

Thanks @phoebe-p for this useful contribution to the PV community!
