+++ examples/sphere_surface_coarse_34sq.py
@@ -1,0 +1,208 @@
+from joblib import Parallel, delayed
+from mpl_toolkits.mplot3d import Axes3D
+import matplotlib.pyplot as plt
+import numpy as np
+import matplotlib.tri as mtri
+from scipy.spatial import ConvexHull
+from rayflare.textures import xyz_texture
+from time import time
+import os
+import seaborn as sns
+
+d_bulk = 1e-10
+# going to get converted to um (multiplied by 1e6)
+
+def points_on_sphere(N, radius, h):
+    """ Generate N evenly distributed points on the unit sphere centered at
+    the origin. Uses the 'Golden Spiral'.
+    Code by Chris Colbert from the numpy-discussion list.
+    """
+    phi = (1 + np.sqrt(5)) / 2 # the golden ratio
+    long_incr = 2*np.pi / phi # how much to increment the longitude
+
+    dz = 2.0 / float(N) # a unit sphere has diameter 2
+    bands = np.arange(N) # each band will have one point placed on it
+    z = bands * dz - 1 + (dz/2) # the height z of each band/point
+    r = np.sqrt(1 - z*z) # project onto xy-plane
+    az = bands * long_incr # azimuthal angle of point modulo 2 pi
+    x = r * np.cos(az)
+    y = r * np.sin(az)
+
+    x = radius * x
+    y = radius * y
+    z = radius * z + h
+    return x, y, z
+
+def average_g(triples):
+    return np.mean([triple[2] for triple in triples])
+
+# fig = plt.figure()
+# ax = fig.add_subplot(111, projection='3d')
+X, Y, Z = points_on_sphere(2**12, radius=0.8, h=0.242)
+
+X = X[Z >= 0]
+Y = Y[Z >= 0]
+Z = Z[Z >= 0]
+
+Triples = np.array(list(zip(X, Y, Z)))
+
+Triples_back = np.array(list(zip(X, Y, -Z - d_bulk)))
+
+hull = ConvexHull(Triples)
+hull_back = ConvexHull(Triples_back)
+triangles = hull.simplices
+triangles_back = hull_back.simplices
+
+colors = np.array([average_g([Triples[idx] for idx in triangle]) for
+                   triangle in triangles_back])
+
+# collec = ax.plot_trisurf(mtri.Triangulation(X, Y, triangles),
+# #         Z, shade=False, cmap=plt.get_cmap('Blues'), array=colors,
+# #         edgecolors='none')
+# # collec.autoscale()
+#
+# plt.show()
+
+[front, back] = xyz_texture(X, Y, Z)
+
+
+front.simplices = triangles
+front.P_0s = front.Points[triangles[:,0]]
+front.P_1s = front.Points[triangles[:,1]]
+front.P_2s = front.Points[triangles[:,2]]
+front.crossP = np.cross(front.P_1s - front.P_0s, front.P_2s - front.P_0s)
+front.size = front.P_0s.shape[0]
+front.zcov = 0
+
+back.simplices = triangles_back
+back.P_0s = back.Points[triangles_back[:,0]]
+back.P_1s = back.Points[triangles_back[:,1]]
+back.P_2s = back.Points[triangles_back[:,2]]
+back.crossP = np.cross(back.P_1s - back.P_0s, back.P_2s - back.P_0s)
+back.size = back.P_0s.shape[0]
+back.zcov = 0
+
+cross_normalized = back.crossP/np.sqrt(np.sum(back.crossP**2,1))[:,None]
+
+flat = np.abs(cross_normalized[:,2]) > 0.9
+bottom_surface = np.all((back.P_0s[:,2] > -0.1, back.P_1s[:,2] > -0.1, back.P_2s[:,2] > -0.1), axis=0)
[…]
