### 2021-09-02 fdbe2bc
fix surface normals of spheres and remaining translation
