+++ rayflare/ray_tracing/rt.py
@@ -690,8 +690,10 @@ def calc_R(n1, n2, theta, pol):
     Rs = np.abs((n1*np.cos(theta)-n2*np.cos(theta_t))/(n1*np.cos(theta)+n2*np.cos(theta_t)))**2
     Rp = np.abs((n1 * np.cos(theta_t) - n2 * np.cos(theta)) / (n1 * np.cos(theta_t) + n2 * np.cos(theta)))**2
     if pol == 's':
+        # print(n1, n2, theta, pol, Rs)
         return Rs
     if pol == 'p':
+        # print(n1, n2, theta, pol, Rp)
         return Rp
     else:
         return (Rs+Rp)/2
@@ -744,20 +746,16 @@ def single_ray_stack(x, y,  nks, alphas, r_a_0, surfaces, widths,
     stop = False
     I = 1
 
-
-
     r_a = r_a_0 + np.array([x, y, 0])
     r_b = np.array([x, y, 0])
 
-    d = 0.1*(r_b - r_a) / np.linalg.norm(r_b - r_a) # direction (unit vector) of ray
+    d = (r_b - r_a) / np.linalg.norm(r_b - r_a) # direction (unit vector) of ray
 
     # want to translate along d so r_a is right below the surface
     n_d = 1.001*r_a[2]/np.abs(d[2])
 
     r_a = r_a + n_d*d
     # print(r_a)
-    r_b = r_a + d
-
 
     # plt.plot([r_a[0], r_b[0]], [r_a[1], r_b[1]], [r_a[2], r_b[2]])
 
@@ -773,21 +771,23 @@ def single_ray_stack(x, y,  nks, alphas, r_a_0, surfaces, widths,
 
     n_interactions = 0
 
-    while not stop:
+    # print('New stack interaction r_a, d', r_a, d)
+    # print(r_a, r_a_0)
 
+    while not stop:
 
         surf = surfaces[surf_index]
 
-        if randomize and (n_passes > 0):
-            h = surf.z_max - surf.z_min + 0.1
-            r_b = [np.random.rand()*surf.Lx, np.random.rand()*surf.Ly, surf.zcov]
-            n_z = np.ceil(abs(h/d[2]))
-            r_a = r_b - n_z*d
-
-        else:
-
-            r_a[0] = r_a[0]-surf.Lx*((r_a[0]+d[0]*(surf.zcov-r_a[2])/d[2])//surf.Lx)
-            r_a[1] = r_a[1]-surf.Ly*((r_a[1]+d[1]*(surf.zcov-r_a[2])/d[2])//surf.Ly)
+        # if randomize and (n_passes > 0):
+        #     h = surf.z_max - surf.z_min + 0.1
+        #     r_b = [np.random.rand()*surf.Lx, np.random.rand()*surf.Ly, surf.zcov]
+        #     n_z = np.ceil(abs(h/d[2]))
+        #     r_a = r_b - n_z*d
+        #
+        # else:
+        #
+        #     r_a[0] = r_a[0]-surf.Lx*((r_a[0]+d[0]*(surf.zcov-r_a[2])/d[2])//surf.Lx)
+        #     r_a[1] = r_a[1]-surf.Ly*((r_a[1]+d[1]*(surf.zcov-r_a[2])/d[2])//surf.Ly)
 
 
         if direction == 1:
@@ -798,9 +798,11 @@ def single_ray_stack(x, y,  nks, alphas, r_a_0, surfaces, widths,
             ni = nks[mat_index-1]
             nj = nks[mat_index]
 
-        res, theta, phi, r_a, d, _, n_interactions = single_interface_check(r_a, d, ni,
+        res, theta, phi, r_a, d, _, n_interactions, side = single_interface_check(r_a, d, ni,
[…]

[…]
