### 2021-09-08 #3460 @EricaEgg
> @EricaEgg do you have any comments /issues so far?

Hi @phoebe-p , thanks for checking in! Sorry for the delay on this. I have been running into problems installing S4 correctly. I was hesitant to open an issue, since it is not an issue with rayflare, itself. However, it has made it impossible to run the examples. If this seems like a relevant issue for the community and that can be solved here, I can post an issue on the rayflare repo.
