### 2021-09-08 qpv-research-group/rayflare#37 @EricaEgg
[JOSS] Installation issues - S4 
I have installed S4 and rayflare as described in the docs. However, when importing functions for rayflare, I get an error.

This is the import I tried:

`from rayflare.matrix_formalism import process_structure, calculate_RAT, get_savepath`

This was the resulting error:
```
ImportError: dlopen(/Users/user/mainenv/lib/python3.9/site-packages/S4.cpython-39-darwin.so, 2): Library not loaded: @rpath/libboost_serialization.dylib
  Referenced from: /Users/user/mainenv/lib/python3.9/site-packages/S4.cpython-39-darwin.so
  Reason: image not found
```

I am using the macOS teriminal and Jupyter notebooks.

Thank you for the help!
