### 2021-09-08 qpv-research-group/rayflare#37 @phoebe-p
Just to check what's going on with the S4 installation, what happens if you just try:

```
import S4
```

If that works, could you try:

```
S = S4.New((1, 0)
