### 2021-09-08 qpv-research-group/rayflare#37 @phoebe-p
(sorry, closing the issue was a wrong click!)

Just to confirm, did you run `make boost` before making S4, and did it seem to execute successfully? Since the issue seems to be related to the boost library.

Also to check what's going on with S4, could you try just (in Python):

```
import S4
```

I guess that will give an error but if not, what happens with:

```
S = S4.New(((1,0),(0,1)), 5)
```

I don't currently have access to a Mac but have successfully installed S4 on it in the past so hopefully we can make it work!
