### 2021-09-08 qpv-research-group/rayflare#37 @phoebe-p
Another thing: RayFlare is designed to still work even when S4 is not installed (if you want to use it on Windows, or if you are simply not interested in the RCWA functionality). When I run the same imports as you in Windows (S4 not installed), I get:

```
from rayflare.matrix_formalism import process_structure, calculate_RAT, get_savepath
WARNING: The RCWA solver will not be available because an S4 installation has not been found.
Backend TkAgg is interactive backend. Turning interactive mode on.
WARNING: The RCWA solver will not be available because an S4 installation has not been found.
```

So it warns me S4 is not installed, rather than throwing an error. I can run the examples which don't involve RCWA as well. It seems because you have attempted to install S4 but something went wrong, you are now in a situation where RayFlare can't recognize you actually don't have the install, so now nothing works anymore. Hopefully we can fix the S4 install so everything works, but I think the first step has to be deleting the version of S4 you currently have from the Python package tree since it's obviously faulty in some way.
