+++ examples/sphere_surface_converged_load.py
@@ -201,7 +201,7 @@ rtstr = rt_structure(textures=[flat_surf, hyperhemi],
 
 options = default_options()
 
-nxs = [50]
+nx = 50
 
 thetas = np.linspace(0, np.pi / 2 - 0.1, 100)
 
@@ -211,83 +211,68 @@ thetas_3 = thetas[40:60]
 thetas_4 = thetas[60:80]
 thetas_5 = thetas[80:100]
 
-thetas = thetas_5
+thetas_min = []
+thetas_max = []
 
-thetas_min = np.int(10*np.round(180*np.min(thetas)/np.pi, 1))
-thetas_max = np.int(10*np.round(180*np.max(thetas)/np.pi, 1))
+for theta_g in [thetas_1, thetas_2, thetas_3, thetas_4, thetas_5]:
 
-pal = sns.color_palette("rocket", len(nxs))
+    thetas_min.append(np.int(10*np.round(180*np.min(theta_g)/np.pi, 1)))
+    thetas_max.append(np.int(10*np.round(180*np.max(theta_g)/np.pi, 1)))
 
-plt.figure()
+T_v = []
+T_t = []
+n_int = []
+theta_dist = []
 
-for i1, nx in enumerate(nxs):
+
+
+for i1 in range(5):
+
+    theta_min = thetas_min[i1]
+    theta_max = thetas_max[i1]
 
     options.wavelengths = np.array([6e-6])
     options.parallel = False
     options.n_rays = nx**2
     options.xlim = 0.05
     options.ylim = 0.05
-
+    
     options.theta = 0.1
     options.nx = nx
     options.ny = nx
 
-    print(options.n_rays)
 
-    minimum_angle = np.pi - np.pi*17.5/180
+    if os.path.isfile('results/sphere_raytrace_2e12_' + str(options.n_rays) + 'rays_' + str(theta_min) + str(theta_max) + '.txt'):
 
-    T_values = np.zeros(len(thetas))
-    T_total = np.zeros(len(thetas))
-    n_interactions = np.zeros(len(thetas))
-    theta_distribution = np.zeros((len(thetas), options.n_rays))
+        T_values = np.loadtxt('results/sphere_raytrace_2e12_' + str(options.n_rays) + 'rays_' + str(theta_min) + str(theta_max) + '.txt')
+        T_total = np.loadtxt('results/sphere_raytrace_totalT_2e12_' + str(options.n_rays) + 'rays_' + str(theta_min) + str(theta_max) + '.txt')
+        n_interactions = np.loadtxt('results/sphere_raytrace_ninter_2e12_' + str(options.n_rays) + 'rays_' + str(theta_min) + str(theta_max) + '.txt')
+        theta_distribution = np.loadtxt('results/sphere_raytrace_thetas_2e12_' + str(options.n_rays) + 'rays_' + str(theta_min) + str(theta_max) + '.txt')
 
-    if os.path.isfile('results/sphere_raytrace_2e12_' + str(options.n_rays) + 'rays_' + str(thetas_min) + str(thetas_max) + '.txt'):
-
-        T_values = np.loadtxt('results/sphere_raytrace_2e12_' + str(options.n_rays) + 'rays_' + str(thetas_min) + str(thetas_max) + '.txt')
-        T_total = np.loadtxt('results/sphere_raytrace_totalT_2e12_' + str(options.n_rays) + 'rays_' + str(thetas_min) + str(thetas_max) + '.txt')
-        n_interactions = np.loadtxt('results/sphere_raytrace_ninter_2e12_' + str(options.n_rays) + 'rays_' + str(thetas_min) + str(thetas_max) + '.txt')
[…]
