### 2021-09-17 qpv-research-group/rayflare#37 @phoebe-p
@EricaEgg any luck deleting the current S4 install/figuring out what was wrong with it?
