### 2021-09-21 #3460 @EricaEgg
I have completed the checklist and support this work. I currently have an [open issue ](https://github.com/qpv-research-group/rayflare/issues/37), but this is related to the installation of S4, which is not required for rayflare. I am able to run all examples that do not perform RCWA calculations/use S4. Therefore I do not think this should prevent the acceptance of this work, but is instead an issue that can be worked out over time. Thank you @phoebe-p!
