### 2021-09-24 #3460 @phoebe-p
Ok, I have made a new release version which is v1.0.1. Zenodo archive DOI with that release is 10.5281/zenodo.5526536
